Build fix.
* wtf/RAMSize.cpp:
(WTF::computeRAMSize):
If you say you are going to return a size_t, actually return a size_t.

Canonical link: https://commits.webkit.org/104868@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@118031 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WTF/wtf/RAMSize.cpp
@@ -45,7 +45,7 @@ static size_t computeRAMSize()
 {
 #if OS(DARWIN)
     int mib[2];
-    int64_t ramSize;
+    size_t ramSize;
     size_t length;
 
     mib[0] = CTL_HW;
